Interval Compare Pane
Now allows you to add intervals and select them and
change colors they will use.

It will maintain the data as a CompareInterval which
also includes a RideFile representation of the interval.
This is stored in Context in a QList as compareIntervals
and can therefore be re-used across all charts.

We now need to think about putting some widgets onto the
Compare 'bar' for turning compare mode on and off and
deleting / reordering / clearing items.

+++ src/IntervalTreeView.cpp
@@ -66,7 +66,7 @@ IntervalTreeView::mimeData (const QList<QTreeWidgetItem *> items) const
 
     // pack data 
     stream << (quint64)(context); // where did this come from?
-    stream << items.count();
+    stream << (int)items.count();
     foreach (QTreeWidgetItem *p, items) {
 
         // convert to one of ours
@@ -75,9 +75,9 @@ IntervalTreeView::mimeData (const QList<QTreeWidgetItem *> items) const
         // serialize
         stream << p->text(0); // name
         stream << (quint64)(i->ride);
-        stream << i->start << i->stop; // start and stop in secs
-        stream << i->startKM << i->stopKM; // start and stop km
-        stream << i->displaySequence;
+        stream << (quint64)i->start << (quint64)i->stop; // start and stop in secs
+        stream << (quint64)i->startKM << (quint64)i->stopKM; // start and stop km
+        stream << (quint64)i->displaySequence;
 
     }
